chore: move lib/types to types
